requirements.txt
ib_insync==0.9.86
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
requests>=2.28.0
python-telegram-bot>=20.0